requirements.txt
paho-mqtt
colorama
munch